chatGPT/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
da-vinci==0.3.0
Django==3.2
flake8==6.0.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
mccabe==0.7.0
multidict==6.0.4
openai==0.27.2
Pillow==9.4.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
shortuuid==1.0.11
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
psycopg2-binary==2.9.5
